TJFE-52 [rename] screens -> home으로 변경
